refactor: move Add vehicle screen to separate folder
